♻️[Refactor] 운영진 스터디 관리 화면 파일 분할 (#1060)
- OperatorStudyManagementView 를 섹션 전환 셸로 축소 (툴바·시트·알럿만 유지)
- 그룹 관리 섹션을 OperatorStudyGroupSection 으로 분리
- 제출 현황 섹션을 OperatorStudySubmissionSection 으로 분리
- 두 섹션이 공유하던 로딩·더보기 스피너를 StudyManagementLoadingView 로 추출
- 빈 상태·에러에서 함께 쓰던 권한 안내를 StudyGroupPermissionGuideView 로 분리
